Propagate exclusion fixture setup errors
Move the temporary fixture generator into test support and keep its exposed
surface narrow for the integration exclusion tests.

Return `anyhow::Result` from the exclusion test driver so setup failures carry
context through the test harness instead of panicking via `expect`.
